Incoporated into dists
- moved distribution and tests to dists and test_dists
